requirements.txt
Flask
Flask-Discord
python-dotenv
flask_sqlalchemy
selenium
webdriver-manager